Update variable types (int->Long, double->BigDecimal) to correspond to a better data representation

+++ src/main/java/io/github/oliviercap/chefduplacard/domain/stock/StockLine.java
@@ -3,23 +3,25 @@ package io.github.oliviercap.chefduplacard.domain.stock;
 import io.github.oliviercap.chefduplacard.domain.food.Aliment;
 import io.github.oliviercap.chefduplacard.domain.unit.Unit;
 
+import java.math.BigDecimal;
+
 public class StockLine {
 
-    private double quantity;
+    private BigDecimal quantity;
     private Aliment aliment;
     private Unit unit;
 
-    public StockLine(double quantity, Aliment aliment, Unit unit) {
+    public StockLine(BigDecimal quantity, Aliment aliment, Unit unit) {
         this.quantity = quantity;
         this.aliment = aliment;
         this.unit = unit;
     }
 
-    public double getQuantity() {
+    public BigDecimal getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(double quantity) {
+    public void setQuantity(BigDecimal quantity) {
         this.quantity = quantity;
     }
